requirements.txt
faiss-gpu==1.7.2
numpy==1.21.2
scikit-learn==1.3.2
scipy==1.10.1
tensorboard==2.14.0
tensorboard-data-server==0.7.2
torch==2.0.0
transformers==4.42.0